+++ NEWS.md
@@ -1,7 +1,9 @@
 # controller 0.0.0.9000
 
 * Adds `control()`, `control_ci()`, and `control_fuzzy()`
+* Adds `control_matches()` for inspecting how matches were made
 * Adds `control_names()`, `control_names_ci()`, and `control_names_fuzzy()` (#7)
 * Adds example dataset `colour_thesaurus`
 * Adds `read_fish()`
 * Adds a `NEWS.md` file to track changes to the package.
+